jinaai-vector/requirements.txt
# 基础 Web 框架
fastapi
uvicorn[standard]
httpx
python-multipart
huggingface_hub

# 图像处理
Pillow

# AI 与深度学习 (指定版本以保证兼容性)
transformers
torch==2.3.1
torchvision==0.18.1

# 数据处理
numpy

einops
timm

sympy==1.13.1